+++ spec/flipper/adapters/active_support_cache_store_spec.rb
@@ -10,7 +10,8 @@ RSpec.describe Flipper::Adapters::ActiveSupportCacheStore do
     Flipper::Adapters::OperationLogger.new(Flipper::Adapters::Memory.new)
   end
   let(:cache) { ActiveSupport::Cache::MemoryStore.new }
-  let(:adapter) { described_class.new(memory_adapter, cache, expires_in: 10.seconds) }
+  let(:write_through) { false }
+  let(:adapter) { described_class.new(memory_adapter, cache, expires_in: 10.seconds, write_through: write_through) }
   let(:flipper) { Flipper.new(adapter) }
 
   subject { adapter }
@@ -22,11 +23,69 @@ RSpec.describe Flipper::Adapters::ActiveSupportCacheStore do
   it_should_behave_like 'a flipper adapter'
 
   describe '#remove' do
-    it 'expires feature' do
-      feature = flipper[:stats]
+    let(:feature) { flipper[:stats] }
+
+    before do
       adapter.get(feature)
       adapter.remove(feature)
-      expect(cache.read(described_class.key_for(feature))).to be(nil)
+    end
+
+    it 'expires feature and deletes the cache' do
+      expect(cache.read(described_class.key_for(feature))).to be_nil
+      expect(feature).not_to be_enabled
+    end
+
+    context 'with write-through caching' do
+      let(:write_through) { true }
+
+      it 'expires feature and writes an empty value to the cache' do
+        expect(cache.read(described_class.key_for(feature))).to be_empty
+        expect(feature).not_to be_enabled
+      end
+    end
+  end
+
+  describe '#enable' do
+    let(:feature) { flipper[:stats] }
+
+    before do
+      adapter.enable(feature, feature.gate(:boolean), flipper.boolean)
+    end
+
+    it 'enables feature and deletes the cache' do
+      expect(cache.read(described_class.key_for(feature))).to be_nil
+      expect(feature).to be_enabled
+    end
+
+    context 'with write-through caching' do
+      let(:write_through) { true }
+
+      it 'expires feature and writes to the cache' do
+        expect(cache.read(described_class.key_for(feature))).to include(boolean: 'true')
+        expect(feature).to be_enabled
+      end
+    end
+  end
+
+  describe '#disable' do
+    let(:feature) { flipper[:stats] }
+
+    before do
+      adapter.disable(feature, feature.gate(:boolean), flipper.boolean)
+    end
+
+    it 'disables feature and deletes the cache' do
+      expect(cache.read(described_class.key_for(feature))).to be_nil
+      expect(feature).not_to be_enabled
+    end
+
+    context 'with write-through caching' do
+      let(:write_through) { true }
+
+      it 'expires feature and writes to the cache' do
+        expect(cache.read(described_class.key_for(feature))).to include(boolean: nil)
+        expect(feature).not_to be_enabled
+      end
     end
   end
 
@@ -39,6 +98,8 @@ RSpec.describe Flipper::Adapters::ActiveSupportCacheStore do
       search.enable
 
       memory_adapter.reset
+      cache.delete(described_class.key_for(search))
+      cache.delete(described_class.key_for(other))
 
       adapter.get(stats)
       expect(cache.read(described_class.key_for(search))).to be(nil)